added carousel plugin
