+++ sweet-spot-app/src/components/carousel-plugin.tsx
@@ -1,0 +1,66 @@
+import * as React from "react";
+import Autoplay from "embla-carousel-autoplay";
+import { ListPlus, LogIn, UserPlus } from "lucide-react";
+
+import { Card, CardContent } from "@/components/ui/card";
+import {
+  Carousel,
+  CarouselContent,
+  CarouselItem,
+  CarouselNext,
+  CarouselPrevious,
+} from "@/components/ui/carousel";
+
+export function CarouselPlugin() {
+  const plugin = React.useRef(
+    Autoplay({ delay: 6000, stopOnInteraction: true })
+  );
+
+  return (
+    <Carousel
+      plugins={[plugin.current]}
+      className="w-full max-w-xs"
+      onMouseEnter={plugin.current.stop}
+      onMouseLeave={plugin.current.reset}
+      opts={{
+        loop: true,
+      }}
+    >
+      <CarouselContent>
+        <CarouselItem>
+          <div className="p-1">
+            <Card>
+              <CardContent className="flex h-48 items-center justify-center p-6 flex-col">
+                <UserPlus size={100} color="#008081" />
+                <p className="font-semibold">
+                  Create an account for your church to be able to add or remove
+                  members
+                </p>
+              </CardContent>
+            </Card>
+          </div>
+        </CarouselItem>
+        <CarouselItem>
+          <div className="p-1">
+            <Card>
+              <CardContent className="flex aspect-square items-center justify-center p-6">
+                <LogIn />
+                <span className="text-4xl font-semibold">numba 2</span>
+              </CardContent>
+            </Card>
+          </div>
+        </CarouselItem>
+        <CarouselItem>
+          <div className="p-1">
+            <Card>
+              <CardContent className="flex aspect-square items-center justify-center p-6">
+                <ListPlus />
+                <span className="text-4xl font-semibold">numba 3</span>
+              </CardContent>
+            </Card>
+          </div>
+        </CarouselItem>
+      </CarouselContent>
+    </Carousel>
+  );
+}
